Fix issue wrong reduction amount if no tax
Fix the issue if you have a reduction cart rule with amount or % of reduction on products/categories/manufacturers and the order has no tax due to country or VATNUMBER entered. The reduction is calculated with tax. So you have a payement error because the total order is different that the real paied.